Revert "[Task] {PROD4POD-1194} Improve error checking (#592)"
This reverts commit b21becad1efff756dfe39e7fffe5a017e98da5fe.

reverting irrelevant commit due to base change
