+++ core/utils/silly-i18n/test/basic.test.js
@@ -1,20 +1,18 @@
-import { determineLanguage, I18n } from "../src/index.js";
-
 import {
+    determineLanguage,
     LanguageError,
     NonExistingSectionError,
     TranslationKeyError,
-} from "../src/errors";
+    I18n,
+} from "../src/index.js";
 
 const LANGUAGE = "foo";
 const FALLBACK_LANGUAGE = "en";
 let i18n;
 
-const translationData = { quux: { bar: "baz" }, options: { opt: "{{opt}}" } };
-
 beforeAll(() => {
     i18n = new I18n(LANGUAGE, {
-        [LANGUAGE]: translationData,
+        [LANGUAGE]: { quux: { bar: "baz" }, options: { opt: "{{opt}}" } },
     });
 });
 
@@ -27,9 +25,6 @@ describe("Test language determination", () => {
 describe("Test basic configuration", () => {
     it("is created correctly", () => {
         expect(i18n).toBeInstanceOf(I18n);
-    });
-    it("Includes all sections", () => {
-        expect(i18n.sections).toStrictEqual(Object.keys(translationData));
     });
     it("Translates correctly", () => {
         expect(i18n.t("quux:bar")).toBe("baz");
