+++ core/utils/silly-i18n/test/section.test.js
@@ -1,14 +1,11 @@
-import { I18n, I18nSection } from "../src/index";
-
-import { TranslationKeyError, NonExistingSectionError } from "../src/errors";
+import { TranslationKeyError, I18n, I18nSection } from "../src/index.js";
 
 const LANGUAGE = "foo";
 let i18n, i18ns;
-const translationData = { quux: { bar: "baz" }, options: { opt: "{{opt}}" } };
 
 beforeAll(() => {
     i18n = new I18n(LANGUAGE, {
-        [LANGUAGE]: translationData,
+        [LANGUAGE]: { quux: { bar: "baz" }, options: { opt: "{{opt}}" } },
     });
     i18ns = new I18nSection(i18n, "quux");
 });
@@ -30,21 +27,6 @@ describe("Test basic configuration", () => {
         expect(thrownError).toBeInstanceOf(TranslationKeyError);
         expect(thrownError.message).toEqual(
             expect.stringMatching(/does not have/)
-        );
-    });
-});
-
-describe("Throws if there's an error", () => {
-    it("Throws the right exception ", () => {
-        let thrownError;
-        try {
-            i18ns = new I18nSection(i18n, "zuul");
-        } catch (error) {
-            thrownError = error;
-        }
-        expect(thrownError).toBeInstanceOf(NonExistingSectionError);
-        expect(thrownError.message).toEqual(
-            expect.stringMatching(/is not included/)
         );
     });
 });
