--- core/utils/silly-i18n/test/test-data/de/numbers.json
@@ -1,5 +1,0 @@
-{
-    "1": "Ein",
-    "2": "Zwei",
-    "3": "Drei"
-}
