--- core/utils/silly-i18n/test/test-data/en/numbers.json
@@ -1,5 +1,0 @@
-{
-    "1": "One",
-    "2": "Two",
-    "3": "Three"
-}
