--- core/utils/silly-i18n/test/test-data/es/numbers.json
@@ -1,5 +1,0 @@
-{
-    "1": "Uno",
-    "2": "Dos",
-    "3": "Tres"
-}
